requirements.txt
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
wandb
opencv-contrib-python
keras
tensorflow
Flask
python-dotenv 
waitress
matplotlib
pillow